backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
django-filter
reportlab>=3.6.0
google-genai>=0.8.5
google.generativeai